moved refactor TestHelper to RegressionEngine
Created RegressionEngine class and moved TestHelper to it.
Updated the Regression tests static diagram.
